git-active: Use 'git merge-base' to find parent tip
When using webrev with a Mercurial repository we use the common ancestor
of the local head and the head in the parent workspace as the basis of
comparison.  The Git equivalent would be to compare against the common
ancestor of the current branch and origin/master, rather than the branch
point as we did previously.