webrev: Rearrange so SCMs create their own pretty headers
Rearrange the code to add PRETTY_CWS and PRETY_PWS, set by each SCM supporting
it to a string to be used as the "pretty" name of the current and parent
workspace, respectively.  At present this is used to avoid duplication of the
"(at <node>)" logic for Mercurial and Git.