Use __attribute__((__gnu_inline__)) on GCC ASM
As a GCC extension, GCC used to define 'extern inline' to mean that a
function was only ever emitted inline and no visible symbol was emitted.

C99, however, made 'extern inline' mean pretty much the reverse of this, and
GCC 4.3 and above in C99 or GNU99 modes follow the standard, so we have to
decorate each definition to indicate we want the non-standard behaviour.