dis: Decode multibyte nops generated by GCC
This decodes 0f 1f (nopl) and 66 0f 1f (nopw), which seem to be the ones
most commonly generated by gcc (determined by disassembling /kernel).
unused opcodes in [0f18, 0f1f) are technically reserved nops, and used
by Intel as "hinting nops", but I haven't found them generated.

It doesn't seem to deal with '66 2e 0f 1f' (nopw %cs:) but I'm not
certain as to why.  This seems to mostly be generated inter-function
however, and so this may not be such a big deal.