check_rtime: Ignore kldap.so's dependency on libc
When building with GCC 4, kldap.so's dependency on libc is unused.  This
is actually not really used either way.  A hack was used when
integrating kerberos to nullify the 'inline' keyword in various kerberos
headers, this leaves a bunch of functions within those headers as normal
static functions.  GCC eliminates them if they are not used (as they are
not in kldap.so), Studio does not.  When present the functions in
question reference symbols from libc.

Yech.