ws: Support git in the most rudimentary way
In keeping with the rest of our support, there's no concept of "parent
workspace" so this is really just a matter of choosing where to put the
protodefs.

In passing fix the classic bug where GNU dirname of no arguments is an
error, and the rest of the script cries.