nfs.d: Split into client/server libraries
When the compilers emit DWARF, they remove type definitions for types not
referenced by the objects being produced (as opposed to when emitting stabs, in
the Studio case, where this doesn't happen).

This means that nfs.d's dependency on 'genunix' is not sufficient to find the
NFS type (of either client or server), and that the server-specific types will
only be present at all on systems with the NFS server both installed and
loaded.

As such, split nfs.d into an nfssrv.d for the server-specifics, have the server
depend on 'nfssrv' and 'nfs.d', and the client depend on 'nfs' rather than
'genunix'.

This should fix the cases where by DTrace is inoperative even after problematic
closed-bins have been stripped