dwarf: Actually allow relocation processing to happen
Well, this is embarrassing...

Due to differences (apparently) between our libelf and what libdwarf
expects, we don't actually define the R_* relocation types for a
platform unless asked to, because the include of sys/elf_foo.h is
guarded.

Because of poor error reporting, this error is not actually reported by
this version of libdwarf, even though attempting to view the .debug_info
section fails, as it contains relocations of "unknown" or non-absolute
offset (because we don't actually understand _any_).

Luckily (read: annoyingly), since we use -fno-dwarf2-indirect-strings to
prevent the old, non-relocatey libdwarf seeing relocations it dislikes,
we seem to have largely gotten away with (read: irritatingly masked) this.