webrev: Detect and handle renames
While gathering the list of _really_ changed files, record the name they
used to have such that we can deal with renames.

Git does not really support renames, and calculates this at runtime also
providing a 'score' indicating the likelyhood of being a rename.  We
currently only treat as renames those situations with a score > 75 (out
of 100).