cpr: Never allow if compiled with GCC.
As the added commentary states, unfortunately the GAS side of the cpr
resume code was never implemented, and this was then sadly masked with a
botched conditional, so the studio code is built with gas, and then does
the wrong thing, crashing immediately upon resume.

Disable it, for the sake of safety.