3172 Xen PV kernel broken when built with open source /usr/lib/cpp
The first check is to see if __i386__ macro has been defined. If it
is - 32-bit file is used. SmartOS and OmniOS both use open source
cpp in /usr/lib. Its source can be found here. 64-bit header file is
never used on the systems above as __i386__ is one of /usr/lib/cpp
built-ins and is defined by default.

__i386 is undefined for 64-bit C compiler and assembler wrapper
invocations and can be used instead of __i386__ to select the right
include file.