6393 zfs receive a full send as a clone Reviewed by: Matthew Ahrens <mahrens@delphix.com> Reviewed by: Prakash Surya <prakash.surya@delphix.com> Reviewed by: Richard Elling <Richard.Elling@RichardElling.com>
This allows to do a full (non-incremental send) and receive it as a
clone of an existing dataset.  It can leverage nopwrite to share blocks
with the origin.  This can be used to change the relationship of
datasets on the target.  For example, maybe on the source you have:

A ---- B ---- C

And you have sent to the target a full of B, and the incremental B->C:

B ---- C

You later realize that you want to have A on the target.  You will have
to do a full send of A, but nopwrite can save you space on the target if
you receive it as a clone of B, assuming that A and B have some blocks
in common:

B ---- C
 \
  A

Closes #23