6459 cstyle doesn't detect opening braces on the same line as function header Reviewed by: Matthew Ahrens <mahrens@delphix.com> Reviewed by: Alex Reece <alex@delphix.com> Reviewed by: Albert Lee <trisk@omniti.com> Reviewed by: Dan McDonald <danmcd@omniti.com>
Our style guidelines require that the opening brace of a function
definition be on the line after the header, but the cstyle script
doesn't check it.

Additionally, if a function has a very long header, it is split onto to
two lines.  These continuation lines are supposed to be indented by four
spaces, not a tab. However, cstyle does not verify that this is the
case.

This change updates the regular expressions to correctly detect these
cases.

Closes #40