6642 testrunner output can be displayed in the wrong order 6643 zfstest should enforce the required privileges before running.
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Jonathan Mackenzie <jonathan.mackenzie@delphix.com>
Reviewed by: Yuri Pankov <yuri.pankov@nexenta.com>

Lines of output that happen in the same hundredth of a second are
printed in alphabetical order, not the order in which they were
received.

Fix is to modify the sorting done on output lines to sort by timestamp,
but leave the lines in the order they arrived.

The script currently requires users to drop privs before running it. It
should just run the runner script with the required privs.

Fix is to always run the tests under ppriv.

Closes #68