7028 avl_destroy_nodes supports emptying, not just destroying, an avl tree
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: Igor Kozhukhov <ikozhukhov@gmail.com>

Currently, the comment above avl_destroy_nodes only describes the
use-case of destroying the tree. However, once avl_destroy_nodes removes
all elements in a tree, it leaves the tree in a consistent state,
allowing the same tree to be used again. The comment should be updated
to indicate that this functionality is supported.

Closes #115