7016 arc_available_memory is not 32-bit safe
Reviewed by: Igor Kozhukhov <ikozhukhov@gmail.com>
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

The problem is that vmem_size() returns size_t result, which is uint32_t
on 32-bit platforms. Attempt to convert result of subtraction to int64_t
won't give negative value, but instead something about 2^^32.

While we are at it, don't let the zio_arena get too full, and add a
tunable to control how full it can get. Note that this should not have
any impact now that we have the fix for 6914, but it's nice to have the
tunable rather than a hard-coded value.

Closes #113