7066 ::dump uses incredibly annoying syntax for specifying length
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Sebastien Roy <sebastien.roy@delphix.com>
Reviewed by: Igor Kozhukhov <ikozhukhov@gmail.com>
Reviewed by: Gordon Ross <gordon.w.ross@gmail.com>
Reviewed by: Garrett D'Amore <garrett@damore.org>

Having to supply an argument after a comma before the ::dump command is
both inconsistent with every other command in mdb (they take arguments
with options like "-x 100") and unbelievably annoying (because you can't
pipe an address to it). We should add a "-l" length argument to ::dump
which tells it how many bytes to dump.

Closes #121