7236 libumem should be able to abort() when an allocation fails
Reviewed by: Alex Reece <alex@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: Basil Crow <basil.crow@delphix.com>
Reviewed by: Robert Mustacchi <rm@joyent.com>

When libumem isn't able to satisfy an allocation request, it can be hard
to debug why that was true unless you manage to get a core dump or other
debug information from that exact instant so you can analyze the
contents of the heap. This adds that feature to libumem through a new
UMEM_DEBUG environment variable argument called "checknull".

Closes #151