7336 vfork and O_CLOEXEC causes zfs_mount EBUSY
Reviewed by: Matt Ahrens <mahrens@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: Robert Mustacchi <rm@joyent.com>

We can run into a problem where we call into zfs_mount, which in turn
calls is_dir_empty, which opens the directory to try and make sure it's
empty. The issue with the current approach is that it holds the
directory open while it traverses it with readdir, which, due to subtle
interaction with the Java JVM, vfork, and exec can cause a tricky race
condition resulting in zfs_mount failures.

The approach to resolving the issue in this patch is to drop the usage
of readdir altogether, and instead rely on the fact that ZFS stores the
number of entries contained in a directory using the st_size field of
the stat structure. Thus, if the directory in question is a ZFS
directory, we can check to see if it's empty by calling stat() and
inspecting the st_size field of structure returned.

Closes #179