7497 prefetch_io.d performance monitoring script has an error
Reviewed by: Dan Kimmel <dan.kimmel@delphix.com>
Reviewed by: Prakash Surya <prakash.surya@delphix.com>

Running the "prefetch_io.d" dtrace script found in the zfs-test suite
would result in the following error:

    dtrace: failed to compile script /opt/zfs-tests/tests/perf/scripts/prefetch_io.d: line 42: aggregation redefined: @pf
    current: @pf = count( )
    previous: @pf = sum( ) : line 39

This is because the "pf" variable is being used for two different types
of aggregators. This patch simply modifies the script to use two different
variables for the two different types of aggregations being collected.

Closes #206