7498 snprintf is incorrectly annotated with PRINTFLIKE1
Reviewed by: Sebastien Roy <sebastien.roy@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Robert Mustacchi <rm@joyent.com>

snprintf() is incorrectly annotated to be printf-like, because the first
argument is indicated to be the format string, when in fact the 3rd
argument is the format string. This annotation is used by lint to detect
improper improper arguments passed to printf (when the format string is
a string literal).

Thankfully, this mistake hasn't caused any errors to be introduced into
the codebase, because the compiler (gcc -Wformat) also checks the
arguments to snprintf (using a different, unknown mechanism to determine
that the 3rd argument is the format string).

Closes #207