7542 zfs_unmount failed with EZFS_UNSHARENFSFAILED
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Dan Kimmel <dan.kimmel@delphix.com>
Reviewed by: Matt Amdur <matt.amdur@delphix.com>

libshare keeps a cached copy of the sharetab listing in memory, which
can become out of date if shares are destroyed or created while leaving
a libzfs handle open. This results in a spurious unmounting failure
when an NFS share exists but isn't in the stale libshare cache. To solve
this, zfs_init_libshare has been modified to always check if libshare
needs to be reinitialized rather than only doing so in a specific
flagged case.

Closes #226