7545 zdb should disable reference tracking
Reviewed by: Dan Kimmel <dan.kimmel@delphix.com>
Reviewed by: Steve Gonczi <steve.gonczi@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

When evicting from the ARC, we manipulate some refcount_t's, e.g.
arcs_size. When using zdb to examine a large amount of data (e.g.
zdb -bb on a large pool with small blocks), the ARC may have a large
number of entries. If reference tracking is enabled, there will be ~1
reference for each block in the ARC. When evicting, we decrement the
refcount and have to search all the references to find the one that we
are removing, which is very slow.

Since zdb is typically used to find problems with the on-disk format,
and not with the code it is running, we should disable reference
tracking in zdb.

Closes #228