7591 New "zloop" command for running ztest indefinitely
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Prakash Surya <prakash.surya@delphix.com>
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: John Kennedy <john.kennedy@delphix.com>

This change adds a new "zloop" command which is used to execute the
ztest command continually in a loop, using random arguments for each
invocation. When any given invocation of ztest terminates (either
successfully, or via a crash), a new invocation of ztest will be
started (again, using random arguments), until the desired runtime for
zloop is reached (or indefinitely, if no timeout value is specified); at
which point zloop will terminate and it's return code will reflect
whether any of the underlying invocations of ztest had crashed.

Additionally, when ztest crashes and produces a core dump, zloop will
copy this dump to a known, but configurable, location such that it can
be manually inspected later. Likewise, the vdev files that ztest uses
are also stored in a known, but configurable, location.

In addition to copying the ztest core dumps to a known location, the
status and stack trace from each dump will be automatically extracted
and inserted into the zloop log file.

And lastly, this new script has been added to the manifest of the
"pkg:/system/file-system/zfs/tests" package.

Closes #234