7180 potential race between zfs_suspend_fs+zfs_resume_fs and zfs_ioc_rename
Reviewed by: Matt Ahrens <mahrens@delphix.com>
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>

zfs_suspend_fs and zfs_resume_fs are called in the open context.
Because of that it's possible that the filesystem (the datatset)
gets renamed while it is suspended.  In that case we won't be able
to find the dataset by its old name.  On the other hand, we keep
a long hold on the dataset, so we can use it to refresh the objset
pointer.

Closes #216