7736 ZFS Performance tests should log FIO summary output
Reviewed by: John Kennedy <john.kennedy@delphix.com>
Reviewed by: Dan Kimmel <dan.kimmel@delphix.com>
Reviewed by: Stephen Blinick <stephen.blinick@delphix.com>

Current zfs performance tests don't store the FIO summary output to
discrete files along-side the output from the monitoring scripts. A
single log file of stdout of all tests is created in the root
directory but it includes numerous other data and no clear
delineation about which test type produced which output.

This is a small change proposed to have each ZFS performance test also
create a 'fio' output log for each run in the:

    test_results/<testid>/perf_data

directory. The output is obtained using the `--output` parameter of FIO.

Closes #266