Archive the build's "nightly.log" file
This change extends the automation to archive the "nightly.log" file
produced by the nightly build. While it's not common, certain build
failures can arise that require this file in order to effectively
root cause the problem. This allows easy access to that file via the
"Artifacts" section of each Jenkins build.

Closes #394