8558 lwp_create() returns EAGAIN on system with more than 80K ZFS filesystems
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Sebastien Roy <sebastien.roy@delphix.com>

On a system with more than 80K ZFS filesystems, we've seen cases where
lwp_create() will start to fail by returning EAGAIN. The problem being,
for each of those 80K ZFS filesystems, a taskq will be created for each
dataset as part of the ZIL for each dataset.

For each of these taskq's, a kernel thread will be created which results
in 24KB being allocated for each thread. With enough of these 24KB
allocations, we eventually exhaust the memory region set aside for these
allocations. Currently, segkpsize is set to a value of 2GB, which means
we can only support about 80K filesystems; 2GB / 24KB = ~80K.

The lwp_create() failure comes into play due to the fact that LWP
creation also allocates 24KB from this same region of memory. Thus, if
we've exhausted this region of memory due to the number of ZIL taskq's,
there won't be any memory avaible to allow the call to lwp_create() to
succeed.

The solution taken by this patch is to modify the taskq used by the ZIL
to be a per-pool taskq, instead of a per-dataset taskq.

Additionally, to increase observability of when we fail to dispatch to
the taskq from within zil_clean(), the taskq internals have also been
slightly reworked to track and expose the "nomem" kstat field for
non-dynamic taskqs; previously this field only existed for dynamic
taskq's. This should allow one to relatively easily determine if
zil_clean() is unable to successfully dispatch, which could impact the
performance of spa_sync().

Closes #439