8565 zpool dumps core trying to destroy unavail/faulted pool and try export it
Reviewed by: Matt Ahrens <mahrens@delphix.com>
Reviewed by: Serapheim Dimitropoulos <serapheim@delphix.com>
Reviewed by: Igor Kozhukhov <igor@dilos.org>

There are cases where sa_get_one_zfs_share() of libshare doens't
fill in all the paths[] due to a specific handle being NULL and
returns SA_SYSTEM_ERR.

Currently regardless of whether this error is returned or not
we always free each of those paths in sa_init_impl() which
can lead to a double-freeing fault.

Since sa_get_one_zfs_share() is also the function that allocates
the memory that will later be pointed by paths[], it should
also be responsible for freeing any no-null elements of that
array in the case of error and also set paths_len to 0 so
the caller doesn't attempt to do anything wrong.

Closes #473