Reboot and verify kernel version after upgrade
This change updates the automation to reboot the system after a nightly
build and upgrade, to verify the running kernel version is what we
expect; i.e. to verify it's running the kernel built via the nightly
build.

This is useful if the upgrade process didn't happen to upgrade the
kernel using the nightly build package, which can happen, and doesn't
even result in an error when this occurs.