illumos sysroot 20181213 v1
In order to produce a sysroot archive for cross compilation, we need to
be able to build binary artefacts from a specific historic version of
the gate.  In order to build the gate on a modern version of the OS, we
must first backport a few minor fixes to the tools and work around some
leakage of build system headers into the gate build.  We also disable a
few components which should not end up in the sysroot.